requirements.txt
requests>=2.31.0
meshtastic>=2.2.0
pytz>=2023.3
python-dotenv>=1.0.0
